% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Anotacion_y_Significacion_Biologica.R
\name{annotated}
\alias{annotated}
\title{Anotación de genes}
\usage{
annotated(listOfTabs, annotPackage, maPackage, ID)
}
\arguments{
\item{listOfTabs}{Lista que contiene los datos obtenidos en el análisis de
expresión diferencial. Cada contraste debe estar almacenado en un elemento
de la lista en forma de marco de datos que contenga los nombres de los
probes (microarrays) o de los genes de Ensembl (RNA-Seq) como nombres de
fila. Los dos últimos elementos de la lista no se usan para la anotación.}

\item{annotPackage}{Paquete de anotaciones para el organismo estudiado.}

\item{maPackage}{Paquete de anotaciones para microarrays.}

\item{ID}{Carácter. Identificador de los genes analizados. Para análisis de
microarray \code{ID = "PROBEID"}, mientras que para análisis de RNA-Seq
\code{ID = "ENSEMBL".}}
}
\value{
Transforma la lista introducida añadiendo las anotaciones de los
genes.
}
\description{
Realiza la anotación de los genes contenidos en un marco de datos, añadiendo
a los datos ya almacendados los identificadores de PROBE (para microarrays) o
ENSEMBL (para RNA-Seq), de ENTREZ, y los símbolos y nombres de los genes.
}
\examples{
# Para datos de microarrays
anot_dea_microarray <- annotated(dea_microarray, maPackage = "hgu133a.db",
                                 ID = "PROBEID")
head(anot_dea_microarray[[1]])

# Para datos de RNA-Seq
anot_dea_RNASeq <- annotated(dea_RNASeq, annotPackage = org.Hs.eg.db,
                             ID = "ENSEMBL")
head(anot_dea_RNASeq[[1]])
}
