% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_Anotacion_y_Significacion_Biologica.R
\name{bsa}
\alias{bsa}
\title{Analisis de significación biológica}
\usage{
bsa(listOfTabs, GOPackage, PATHPackage, organism, pvcoff = 0.05, padmethod)
}
\arguments{
\item{listOfTabs}{Lista resultante de realizar las anotaciones.}

\item{GOPackage}{Mapas entre los ID de genes Entrez y los ID de Gene Ontology
(GO)}

\item{PATHPackage}{Mapeos entre identificadores de genes Entrez e
identificadores de vías KEGG}

\item{organism}{Carácter. Organismo sobre el que se realiza el análisis.}

\item{pvcoff}{Numérico. Umbral de corte del p-valor para filtrar los
resultados.}

\item{padmethod}{Carácter. Método utilizado para ajustar los p-valores para
pruebas múltiples. Consulte \code{\link[stats]{p.adjust}} para ver la lista
completa de opciones.}
}
\value{
Una lista con tantos elementos como marcos de datos de genes contenía
la entrada. Cada elemento de la lista es un objeto de la clase
\code{enrichResult}.
}
\description{
Realiza el análisis de enriquecimiento mediante el paquete ReactomePA
}
\examples{
# Para datos de microarrays
enRes_microarray <- bsa(anot_dea_microarray, GOPackage = "org.Hs.egGO",
                        PATHPackage = "org.Hs.egPATH", organism = "human",
                        pvcoff = 0.05, padmethod = "BH")
enRes_microarray

# Para datos de RNA-Seq
enRes_RNASeq <- bsa(anot_dea_RNASeq, GOPackage = "org.Hs.egGO",
                    PATHPackage = "org.Hs.egPATH", organism = "human",
                    pvcoff = 0.05, padmethod = "BH")
enRes_RNASeq
}
