% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graficos.R
\name{col_vect}
\alias{col_vect}
\title{Vector de colores}
\usage{
col_vect(object)
}
\arguments{
\item{object}{Objeto.}
}
\value{
vector de caracteres (colores).
}
\description{
Genera un vector de colores por muestra para objetos
\code{ExpressionFeatureSet}, \code{ExpressionSet}, \code{DGEList},
\code{OnDiskMSnExp}, \code{MSnExp}, \code{SummarizedExperiment} y
\code{MSnSet}.
}
\examples{
col_vect(data_microarray)
}
