% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_Analisis_Expresion_Diferencial.R
\name{dea}
\alias{dea}
\title{Análisis de Expresión Diferencial}
\usage{
dea(
  object,
  cont = NULL,
  name = NULL,
  maxanal = NULL,
  adjmethod = "BH",
  pvalcoff = NULL,
  dtmethod = "separate"
)
}
\arguments{
\item{object}{Objeto de la clase \code{ExpressionSet}}

\item{cont}{Carácter, especifica los contrastes a realizar. Los nombres de
los grupos experimentales se deben introducir de idéntica manera a como
están codificados en el objeto, separados por un guión.}

\item{name}{Carácter, especifica nombres alternativos para los contrastes. Si
\code{name = NULL} se usaran nombres iguales a los usados en el parámetro
\code{cont}.}

\item{maxanal}{Numérico. Número máximo de analitos a mostrar en el resultado.}

\item{adjmethod}{Carácter. Método utilizado para ajustar los p-valores para pruebas
múltiples. Consulte \code{\link[stats]{p.adjust}} para ver la lista
completa de opciones.}

\item{pvalcoff}{Numérico. Se filtrarán todos los analitos con un p-valor
mayor al especificado.}

\item{dtmethod}{Cadena de caracteres que especifica cómo se combinarán los
genes y los contrastes en el esquema de prueba múltiple. Las opciones son
\code{"separate"}, \code{"global"}, \code{"hierarchical"} o
\code{"nestedF"}.}
}
\value{
Una lista con un marco de datos para cada contraste realizado, una
tabla que muestra la cantidad de analitos sobreexpresados e infraexpresados
por contraste y un objeto \code{TestResults} que mantiene los analitos que
han mostrado expresión diferencial en algún contraste. Para cada contraste
se muestra un 1 si está sobreexpresado u -1 si está infraexpresado y un 0
si no hay expresión diferencial. Los marcos de datos de cada contraste
muestran diferente estadísticos obtenidos en el análisis (diferencia media
(logFC), expresión promedio (AveExpr), estadístico t moderado, p-valor,
p-valor ajustado y estadístico B), y se muestra ordenado por p-valor de
forma ascendente.
}
\description{
Realiza un análisis de expresión diferencial basado en modelos lineales del
paquete limma.
}
\examples{
# Para los datos de microarrays.
dea_microarray <- dea(norm_data_microarray, cont = "Treated-Untreated",
                      name = "TRvsUN", maxanal = 1000, adjmethod = "fdr",
                      pvalcoff = 0.1, dtmethod = "separate")
dea_microarray

# Para los datos de RNASeq.
dea_RNASeq <- dea(norm_data_RNASeq, cont = "BCell-Kidney", name = "BCvsKi",
                  adjmethod = "BH", pvalcoff = 0.1, dtmethod = "global")
dea_RNASeq

# Para los datos de GC/LC-MS RS.
dea_MetabRS <- dea(norm_data_MetabRS, cont = "CD-Control", name = "CDvsCon",
                   adjmethod = "holm", pvalcoff = 0.1,
                   dtmethod = "hierarchical")
dea_MetabRS

# Para los datos de contendedores de espectros de MS/NMR (se crea un nuevo
# grupo falso para mostrar el funcionamiento con 3 grupos experimentales).
newgroups <- c(rep("patient", 15), rep("treated", 15), rep("control", 17))
pData(norm_data_MetabSB)[1] <- newgroups
dea_MetabSB <- dea(norm_data_MetabSB, cont = c("patient-treated",
                   "patient-control", "treated-control"),
                   name = c("PAvsTR", "PAvsCO", "TRvsCO"),
                   adjmethod = "bonferroni", pvalcoff = 0.1,
                   dtmethod = "nestedF")
dea_MetabSB

# Para los datos de concentraciones de metabolitos (se crean nuevos grupos
# falsos para mostrar el funcionamiento de la intersección).
newgroups <- c(rep("cac.m", 21), rep("cac.w", 20),
               rep("con.m", 15), rep("con.w", 15))
pData(norm_data_MetabMC)[1] <- newgroups
dea_MetabMC <- dea(norm_data_MetabMC, cont = c("cac.m-con.m", "cac.w-con.w",
                   "(cac.m-con.m)-(cac.w-con.w)"), name = c("CACvsCON.M",
                   "CACvsCON.W", "INT"), adjmethod = "BY", pvalcoff = 0.1,
                   dtmethod = "separate")
dea_MetabMC
}
