% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Instalaci칩n_paquetes.R
\name{ins_pack}
\alias{ins_pack}
\title{Instalaci칩n de paquetes de Bioconductor}
\usage{
ins_pack(annot_pack = NULL)
}
\arguments{
\item{annot_pack}{Paquete de anotaciones para los datos del an치lisis.}
}
\value{

}
\description{
Instalaci칩n y carga los paquetes de Bioconductor necesarios para el correcto
funcionamiento del paquete en caso de que no lo estuvieran
}
\examples{
ins_pack()
}
