% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_gráficos.R
\name{ngroup}
\alias{ngroup}
\title{Grupos experimentales}
\usage{
ngroup(object)
}
\arguments{
\item{object}{Objeto.}
}
\value{
Numérico (número de grupos experimentales).
}
\description{
Da el Número de grupos experimentales de objetos \code{ExpressionFeatureSet},
\code{DGEList}, \code{OnDiskMSnExp}, \code{MSnExp},
\code{SummarizedExperiment} y \code{MSnSet}.
}
\examples{
ngroup(data_microarray)
}
