% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_graficos.R
\name{plotTFM}
\alias{plotTFM}
\title{Gráficos}
\usage{
plotTFM(object, plot)
}
\arguments{
\item{object}{Objeto del tipo \code{ExpressionFeatureSet},
\code{ExpressionSet}, \code{DGEList}, \code{OnDiskMSnExp}, \code{MSnExp},
\code{SummarizedExperiment} o \code{MSnSet}.}

\item{plot}{Tipo de gráfico. Puede ser \code{"PCA"} para gráfico de
componentes principales, \code{"boxplot"} para diagramas de caja,
\code{"heatmap"} para un mapa de calor de la varianza de genes,
\code{"barplot"} para un gráfico de barras de las librerías usadas (sólo
para datos de RNA-Seq), \code{"BPC"} para un cromatograma de pico base
(sólo para espectros de GC/LC-MS), \code{"heatmapS"} para un mapa de calor
por muestras o \code{"density"} para un gráfico de densidad por muestra.}
}
\value{
Un gráfico distinto en función de la selección realizada.
}
\description{
Realiza diferentes tipos de gráficos en función del objeto introducido.
Acepta objetos del tipo \code{ExpressionFeatureSet}, \code{ExpressionSet},
\code{DGEList}, \code{OnDiskMSnExp}, \code{MSnExp},
\code{SummarizedExperiment} o \code{MSnSet}.
}
\examples{
plotTFM(object, plot = c("PCA", "boxplot", "heatmap", "barplot", "BPC",
                         "heatmapS", "density")
}
