% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Preprocesamiento_y_normalización.R
\name{prep_metRS}
\alias{prep_metRS}
\title{Preprocesamiento de espectros brutos de GC/LC-MS}
\usage{
prep_metRS(
  object,
  fCPmethod = NULL,
  refineRT = FALSE,
  refineIn = FALSE,
  refineMN = FALSE,
  aRtmethod = NULL,
  gCPmethod = NULL,
  fillCP = FALSE,
  fillmethod = NULL,
  ...
)
}
\arguments{
\item{object}{Objeto del tipo \code{OnDiskMSnExp} o \code{MSnExp}.}

\item{refineRT, refineIn, refineMN}{Lógico, si es \code{TRUE} se aplica un
filtraje según un determinado tiempo de retención (\code{refineRT}) o
intensidad (\code{refineIn}), o se fusionan picos cromatográficos que se
superponen o presentan unas dimensiones de rt y m/z cercanas
(\code{refineMN}).}

\item{aRtmethod}{Método de alineamiento de picos a usar. Para más información
visite \code{\link[xcms]{adjustRtime}}.}

\item{gCPmethod}{Método de correspondencia a usar. Para más información
visite \code{\link[xcms]{groupChromPeaks}}.}

\item{fillCP}{Lógico, si es \code{TRUE} se realiza un rellenado de los
valores ausentes.}

\item{fillmethod}{Método de rellenado de valores ausentes a usar. Para más
información visite \code{\link[xcms]{fillChromPeaks}}.}

\item{...}{Más argumentos que se pasan a \code{\link[xcms]{findChromPeaks}} o
sus métodos, \code{\link[xcms]{refineChromPeaks}} o sus métodos,
\code{\link[xcms]{adjustRtime}} o sus métodos,
\code{\link[xcms]{groupChromPeaks}} o sus métodos, o
\code{\link[xcms]{fillChromPeaks}} o sus métodos.}

\item{fcPmethod}{Método de deteccion de picos a usar. Para más información
visite \code{\link[xcms]{findChromPeaks}}.}
}
\value{
Un objeto de la clase \code{MSnSet} que contiene los datos
resultantes del preprocesamiento aplicado
}
\description{
Realiza el preprocesamiento de espectros brutos de GC/LC-MS almacenados en
objetos del tipo \code{OnDiskMSnExp} o \code{MSnExp}.
}
\examples{
data_MetabRS <- prep_metRS(data_MetabRS,
                           fCPmethod = c("centWave", "centWavewpi",
                           "matchedFilter", "massifquant", "MSW"),
                           refineRT = TRUE, refineIn = TRUE,
                           refineMN = TRUE,
                           aRtmethod = c("peakGroups", "obiwarp"),
                           gCPmethod = c("density", "mzClust", "nearest"),
                           fillCP = TRUE, fillmethod = c("fill", "area"))
data_MetabRS
}
