% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Preprocesamiento_y_normalizacion.R
\name{prep_metSB}
\alias{prep_metSB}
\title{Preprocesamiento de contenedores de espectros de MS o NMR}
\usage{
prep_metSB(
  object,
  filterMV = TRUE,
  filterF = TRUE,
  filterRSD = TRUE,
  filterB = TRUE,
  mpmv = NULL,
  mf = NULL,
  fFmethod = NULL,
  QClab = NULL,
  mrsd = NULL,
  mfc = NULL,
  Blab = NULL,
  remB = TRUE,
  fib = 0
)
}
\arguments{
\item{object}{Objeto del tipo \code{SummarizedExperiment}.}

\item{filterMV, filterF, filterRSD, filterB}{Lógico, si es \code{TRUE} se
realiza el filtraje de los datos. Para \code{filterMV} se filtrarán muestra
en función del porcentaje de valores perdidos, para \code{filterF} se
filtrarán características en función del porcentaje de muestras que
contengan valores no perdidos, para \code{filterRSD} se filtrarán
características en función de la desviación estándar relativa de las
muestras de control de calidad. Para \code{filterB} se filtrarán
características de origen no biológico mediante muestras en blanco.}

\item{mpmv}{Numérico, valor entre 0 y 1 del umbral del porcentaje de valores
perdidos en la muestra.}

\item{mf}{Numérico, valor entre 0 y 1 del umbral de fracción de detección.}

\item{fFmethod}{Método del filtraje de características a usar. Para más
información visite \code{\link[pmp]{filter_peaks_by_fraction}}.}

\item{QClab}{Carácter, etiqueta de clase utilizada para identificar muestras
de control de calidad. Para \code{filterB = TRUE}, Si el parámetro QClab no
es NULL, se utilizarán muestras de control de calidad para calcular la
intensidad de la señal media.}

\item{mrsd}{Numérico, valor umbral del porcentaje de RSD del control de
calidad.}

\item{mfc}{Numérico, fold change mínimo entre muestras analíticas y en
blanco.}

\item{Blab}{Carácter, etiqueta de clase utilizada para identificar muestras
en blanco.}

\item{remB}{Lógico, si es TRUE se eliminarán las muestras en blanco.}

\item{fib}{Numérico, valor entre 0 y 1 de la fracción de picos en blanco en
las que deben estar presentes.}
}
\value{
Un objeto de la clase \code{MSnSet} que contiene los datos
resultantes del preprocesamiento aplicado
}
\description{
Realiza el preprocesamiento de contenedores de espectros de MS o NMR
almacenados en objetos del tipo \code{SummarizedExperiment}.
}
\examples{
prep_data_MetabSB <- prep_metSB(data_MetabSB, filterMV = TRUE, mpmv = 0.1,
                                filterF = TRUE, mf = 0.9, fFmethod = "across",
                                filterRSD = FALSE, filterB = FALSE)
prep_data_MetabSB
}
