% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Preprocesamiento_y_normalizacion.R
\name{prep_norm_microarray}
\alias{prep_norm_microarray}
\title{Preprocesamiento y Normalización de datos de microarrays}
\usage{
prep_norm_microarray(
  object,
  varFilter = TRUE,
  featureFilter = TRUE,
  annot_pack = NULL,
  ...
)
}
\arguments{
\item{object}{Objeto del tipo \code{ExpressionFeatureSet}.}

\item{varFilter, featureFilter}{Lógico, si es \code{TRUE} se aplicará el
filtrado por varianza (\code{varFilter}), o el filtrado basado en
anotaciones y la eliminación de duplicados (\code{featureFilter}). Si ambos
son \code{TRUE} se aplican ambos filtrados.}

\item{annot_pack}{Paquete de anotaciones (para datos de microarrays).}

\item{...}{Más argumentos que se pasan a \code{\link[oligo]{rma}} o
\code{\link[genefilter]{nsFilter}}.}
}
\value{
Un objeto de la clase \code{ExpressionSet} que contiene los datos
resultantes del preprocesamiento aplicado
}
\description{
Realiza el preprocesamiento y la normalización de datos de microarrays
almacenados en objetos del tipo \code{ExpressionFeatureSet}.
}
\examples{
norm_data_microarray <- prep_norm_microarray(data_microarray,
                                             varFilter = TRUE,
                                             featureFilter = TRUE,
                                             annot_pack = "hgu133a.db")
norm_data_microarray
}
