% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_Preprocesamiento_y_normalizacion.R
\name{prep_norm_rna}
\alias{prep_norm_rna}
\title{Preprocesamiento y Normalización de datos de RNA-Seq}
\usage{
prep_norm_rna(object, usFilter = TRUE, ...)
}
\arguments{
\item{object}{Objeto del tipo \code{DGEList}.}

\item{usFilter}{Lógico, si es \code{TRUE} se aplicará un filtrado no
específico eliminando filas para las que la mitad de las muestras tengan un
recuento por millón menor a 1.}

\item{...}{Más argumentos que se pasan a \code{\link[edgeR]{calcNormFactors}}}
}
\value{
Un objeto de la clase \code{DGEList} que contiene los datos
resultantes del preprocesamiento aplicado
}
\description{
Realiza el preprocesamiento y la normalización de datos de RNA-Seq
almacenados en objetos del tipo \code{DGEList}.
}
\examples{
norm_data_RNASeq <- prep_norm_rna(data_RNASeq, usFilter = TRUE,
                                  method = "TMM")
norm_data_RNASeq
}
