% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_lectura_de_datos.R
\name{read_data}
\alias{read_data}
\title{Lectura de archivos de datos ómicos}
\usage{
read_data(
  data_type,
  path,
  gz_file = FALSE,
  file_type,
  raw_data,
  sep_rd = "",
  targets,
  sep_targ = "",
  mode = "onDisk"
)
}
\arguments{
\item{data_type}{Tipo de datos a introducir, puede ser \code{"microarray"},
\code{"RNA-Seq"}, \code{"MetabRS"}, \code{"MetabSB"} o \code{"MetabMC"}.}

\item{path}{Ruta de ubicación de los archivos para \code{data_type =
"microarray"} o \code{"MetabRS"}.}

\item{gz_file}{Lógico, si es \code{TRUE} se leerán archivos con extensión
".CEL.gz")}

\item{file_type}{Tipo de archivos, para \code{data_type = "MetabRS"}. Puede
ser \code{".NetCDF"}, \code{".mzML"}, \code{".mzXML"} o \code{".mzData"}.}

\item{raw_data}{Ruta y nombre del archivo con los datos brutos del
experimento, para \code{data_type = "RNA-Seq"}, \code{"MetabSB"} o
{"MetabMC"}. Debe ser un archivo \code{.csv} o \code{.txt} con cabecera. La
primera columna debe corresponder a los genes (\code{"RNA-Seq"}), los bins
(\code{"MetabSB"}) o los metabolitos (\code{"MetabMC"}) y las siguientes a
las muestras.}

\item{sep_rd}{Carácter separador de campo para el archivo \code{raw_data}.
Los valores de cada línea del archivo están separados por este carácter. De
forma predeterminada \code{sep_rd = ""}.}

\item{targets}{Ruta y nombre del archivo con la información sobre las
muestras del experimento. Debe ser un archivo \code{.csv} o \code{.txt} con
cabecera. Para \code{data_type = "microarray"} o \code{"MetabRS"} debe
tener tres columnas como mínimo, la primera con el nombre de los archivos
que contienen los datos de cada muestra, la segunda con un identificador
único de las muestras y la tercera con el grupo experimental de las
muestras. Para el resto debe tener dos columnas como mínimo, la primera con
el identificador de las muestras y la segunda con el grupo experimental de
las muestras. En ambos casos se aceptan más columnas como covariables.}

\item{sep_targ}{Carácter separador de campo para el archivo \code{targets}.
Los valores de cada línea del archivo están separados por este carácter. De
forma predeterminada \code{sep_targ = ""}.}

\item{mode}{Puede ser \code{"inMemory"} (predeterminado) o \code{"onDisk"}.
El primero carga los datos sin procesar en la memoria, mientras que el
segundo solo genera el objeto y se accede a los datos sin procesar en el
disco cuando es necesario (sólo válido para \code{data_type = "MetabRS"}).}
}
\value{
Un objeto de la clase \code{ExpressionFeatureSet} (para microarrays),
\code{DGEList} (para RNA-Seq), \code{OnDiskMSnExp} o \code{MSnExp} (para
espectros brutos de GC/LC-MS), \code{SummarizedExperiment} (para
contenedores de espectros de MS/NMR), o \code{MSnSet} (para concentraciones
de metabolitos) que contiene los datos de los ensayos y datos sobre las
muestras.
}
\description{
Lectura de diferentes tipos de archivos de datos ómicos y construcción de un
objeto en el que almacenar los datos de los ensayos y las muestras.
}
\examples{
# Para microarrays
data_microarray <- read_data(data_type = "microarray",
                             path = "./data/microarray",
                             targets = "./data/microarray/targets.csv",
                             sep_targ = ";")
data_microarray

# Para RNA-Seq
data_RNASeq <- read_data(data_type = "RNA-Seq",
                         raw_data = "./data/RNA-Seq/counts.csv",
                         sep_rd = "\t",
                         targets = "./data/RNA-Seq/targets.csv",
                         sep_targ = ",")
data_RNASeq

# Para Espectros Brutos de GC/LC-MS
data_MetabRS <- read_data(data_type = "MetabRS",
                          path = "./data/Met_mzML",
                          file_type = ".mzML",
                          targets = "./data/Met_mzML/targets.csv",
                          sep_targ = ";",
                          mode = "onDisk")
data_MetabRS

# Para Contenedores de Espectros de MS/NMR
data_MetabSB <- read_data(data_type = "MetabSB",
                          raw_data = "./data/Met_spectra_bins/spectra.csv",
                          sep_rd = ",",
                          targets = "./data/Met_spectra_bins/targets.csv",
                          sep_targ = ",")
data_MetabSB

# Para Concentraciones de Metabolitos
data_MetabMC <- read_data(data_type = "MetabMC",
                         raw_data = "./data/Met_conc/concent.csv",
                         sep_rd = ",",
                         targets = "./data/Met_conc/targets.csv",
                         sep_targ = ",")
data_MetabMC
}
