% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_Instalacion_paquetes.R
\name{ins_pack}
\alias{ins_pack}
\title{Instalación de paquetes de Bioconductor}
\usage{
ins_pack(annot_pack = NULL)
}
\arguments{
\item{annot_pack}{Paquetes de anotaciones para los datos del análisis.}
}
\value{

}
\description{
Instala y carga los paquetes de Bioconductor necesarios para el correcto
funcionamiento del paquete en caso de que no lo estuvieran.
}
\examples{
ins_pack()
}
